initial markings + classifier bug fix
added initial markings as an attribute of Place (default 0).
changed the visualizer to accept this as a starting value.
added a reset button to the simulator to reset all the markings to their init values.

fixed a bug in the classifier where it was using + for array concat.
this caused a singleton + empty to be just the singleton element (not even an array anymore).